test: convert e2e asset scoping tests to local server (#183)
Convert 19 disabled Playwright tests from tests/e2e/ (hardcoded to Azure)
into tests_e2e/test_ui_asset_scoping.py using local uvicorn server.

Covers:
- TestUIAssetVisibility (5): group-based visibility, admin sees all, badges
- TestUIGroupPickerUpload (6): upload panel group picker interactions
- TestUIGroupPickerLibrary (7): library detail-row group picker
- TestSchedulerGroupScoping (1): scheduler target dropdown scoping

The 10 API-only tests are already covered by tests/test_rbac.py and were
not duplicated. Deletes the old tests/e2e/ directory entirely.
